5.0.0: a published file's size needs 64 bits
resproxy_ip_14d's MMDB is 4.58 GB and the spec typed every byte count as a bare
`integer`, so the generator emitted Integer and database().list() threw for
every max-tier caller - not for that one entry, for the whole catalogue. Swift
and Zig were hand-written and both chose 64-bit, which is the tell that the spec
was the defect rather than the generators.

MAJOR because the type of a public accessor changed, even though the call path
it is on could not succeed before.

The integration suite moves with it, and also to the CURRENT vocabulary: it was
pinned to [1.0,2.0), so the dataset->database rename never compiled against it
and its red read as a broken staging API for four majors.

+++ src/main/java/io/vpndetection/model/DatabaseFormatSize.java
@@ -45,7 +45,7 @@ public class DatabaseFormatSize {
 
   public static final String JSON_PROPERTY_BYTES = "bytes";
   @javax.annotation.Nullable
-  private Integer bytes;
+  private Long bytes;
 
   public DatabaseFormatSize() { 
   }
@@ -74,26 +74,26 @@ public class DatabaseFormatSize {
   }
 
 
-  public DatabaseFormatSize bytes(@javax.annotation.Nullable Integer bytes) {
+  public DatabaseFormatSize bytes(@javax.annotation.Nullable Long bytes) {
     this.bytes = bytes;
     return this;
   }
 
   /**
-   * Size of the published file, or null when it has not been published yet
+   * Size of the published file, or null when it has not been published yet. int64 because it is not hypothetical: resproxy_ip_14d&#39;s MMDB is 4.58 GB, so a 32-bit field cannot carry the catalogue and &#x60;list&#x60; throws for every caller rather than for that one entry. 
    * @return bytes
    */
   @javax.annotation.Nullable
   @JsonProperty(value = JSON_PROPERTY_BYTES, required = false)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public Integer getBytes() {
+  public Long getBytes() {
     return bytes;
   }
 
 
   @JsonProperty(value = JSON_PROPERTY_BYTES, required = false)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public void setBytes(@javax.annotation.Nullable Integer bytes) {
+  public void setBytes(@javax.annotation.Nullable Long bytes) {
     this.bytes = bytes;
   }
